\name{BCRANK-internal}
\docType{methods}
\alias{bcrankScore,BCRANKmatch-method}
\alias{consensus,BCRANKmatch-method}
\alias{matchVector,BCRANKmatch-method}
\alias{show,BCRANKmatch-method}
\alias{pwm,BCRANKsearch-method}
\alias{searchPath,BCRANKsearch-method}
\alias{show,BCRANKsearch-method}
\alias{fname,BCRANKresult-method}
\alias{toptable,BCRANKresult-method}
\alias{show,BCRANKresult-method}

\title{BCRANK-internal}
\description{Internal methods for the \pkg{BCRANK} package.}
\seealso{
  \code{\link[BCRANK:bcrank]{bcrank}}
}

\author{Adam Ameur, \email{adam.ameur@lcb.uu.se}}

\keyword{methods}
\keyword{internal}
