\name{BCRANKout}
\alias{BCRANKout}
\docType{data}
\title{BCRANK results for USF1 ChIP-chip data}
\description{
  Results from running \code{bcrank} on USF1 whole genome ChIP-chip
  data for the human liver cell line HepG2. 
}
\usage{data(BCRANKout)}
\source{
  Data from whole genome ChIP-chip experiments on human liver cell line
  HepG2. (Rada-Iglesias, A., et \emph{al}. 2007)
}
\references{
  Rada-Iglesias, A., et \emph{al}. (2007) Whole-genome maps of USF1 and
  USF2 binding and histone H3 acetylation reveal new aspects of promoter
  structure and candidate genes for common human disorders.\emph{Genome
	Research}, Accepted
}
\seealso{
  \code{\link[BCRANK:bcrank]{bcrank}}
}

\keyword{datasets}
